requirements.txt
pytest
requests
jsonschema
pydantic
pytest-pyreport
python-dotenv